fix: remove unused files and rename CustomLink.tsx
