+++ apps/web/src/app/page.tsx
@@ -4,9 +4,6 @@ import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
 import Link from '@mui/material/Link';
 import NextLink from 'next/link';
-import ProTip from '../components/PropTip';
-import Copyright from '../components/Copyright';
-
 
 export default function Home() {
   return (
@@ -23,11 +20,9 @@ export default function Home() {
         <Typography variant="h4" component="h1" sx={{ mb: 2 }}>
           Material UI - Next.js App Router example in TypeScript
         </Typography>
-        <Link href="/about" color="secondary" component={NextLink}>
-          Go to the about page
+        <Link href="/login" color="secondary" component={NextLink}>
+          Go to the login page
         </Link>
-        <ProTip />
-        <Copyright />
       </Box>
     </Container>
   );
